# Preprocess input before parsing