--- a/json/codec.go
+++ b/json/codec.go
@@ -20,7 +20,10 @@ type codec struct {
 }
 
 type encoder struct{ flags AppendFlags }
-type decoder struct{ flags ParseFlags }
+type decoder struct {
+	flags      ParseFlags
+	inputFlags inputFlags
+}
 
 type encodeFunc func(encoder, []byte, unsafe.Pointer) ([]byte, error)
 type decodeFunc func(decoder, []byte, unsafe.Pointer) ([]byte, error)
@@ -983,11 +986,11 @@ func syntaxError(b []byte, msg string, args ...interface{}) error {
 	return e
 }
 
-func inputError(b []byte, t reflect.Type) ([]byte, error) {
+func inputError(b []byte, t reflect.Type, flags inputFlags) ([]byte, error) {
 	if len(b) == 0 {
 		return nil, unexpectedEOF(b)
 	}
-	_, r, err := parseValue(b)
+	_, r, err := parseValue(b, flags)
 	if err != nil {
 		return r, err
 	}