This PR updates `json.Parse()` to first look at the string to determine whether it's valid ASCII print and whether it contains a backslash. This information is then made available to the parsing routines, allowing them to take fast paths without having to check each of their inputs independently.

```
name           old time/op    new time/op    delta
CodeDecoder-4    6.71ms ± 3%    5.49ms ± 1%  -18.18%  (p=0.008 n=5+5)

name           old speed      new speed      delta
CodeDecoder-4   289MB/s ± 3%   354MB/s ± 1%  +22.19%  (p=0.008 n=5+5)
```

Preprocessing the input is quick thanks to SIMD optimized routines (https://github.com/segmentio/encoding/pull/66 and `bytes.IndexByte`). There are perhaps further improvements to be made by coalescing the passes into one.